data-loaders/requirements/requirements.txt
# Database connectivity
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# CLI and utilities
click>=8.1.0
tqdm>=4.66.0

# Logging and monitoring
rich>=13.5.0

# Optional: For advanced data validation
cerberus>=1.3.4
